requirement.txt
aiohttp
python-binance